+++ src/commons.ts
@@ -1,12 +1,18 @@
+import * as fs from "fs";
+
 import * as vscode from "vscode";
 import * as ps from "child_process";
 
+import * as YAML from "yaml";
+
 const detect = require("detect-port");
 
+type Service = {
+  port: number | null;
+};
+
 export function isForgeRunning(): boolean {
-  const stdout = ps
-    .execSync("ps aux | grep -e forge -e target/app -e runserver -e uvicorn")
-    .toString();
+  const stdout = ps.execSync("ps aux | grep -w 'forge run'").toString();
   const processes = stdout
     .split("\n")
     .filter((p) => !p.includes("grep") && p.trim().length > 0);
@@ -32,25 +38,28 @@ export function getMindsmithsTerminal(): [vscode.Terminal, boolean] {
 
 export async function updatePortsStatusBarItem(
   item: vscode.StatusBarItem,
+  projectLocalConfig: any,
   cycles: number = 20
 ) {
   if (item.text === null || item.text === undefined) {
     item.text = "No active ports";
   }
 
-  const listedPorts = getListedPorts();
-  if (listedPorts.length > 0) {
+  const listedPorts = getListedPorts(projectLocalConfig);
+  if (listedPorts.size > 0) {
     let takenPorts: Array<Number> = [];
-    listedPorts.forEach((port) => {
+    for (const [port, serviceName] of listedPorts) {
       detect(port).then((matched: number) => {
         if (port !== matched) {
           takenPorts.push(port);
           takenPorts.sort();
         }
         if (takenPorts.length > 0) {
-          item.text = "Ports: " + takenPorts.join(", ");
+          item.text =
+            "Ports: " +
+            takenPorts.map((port) => listedPorts.get(port)).join(", ");
           const oldTooltip = item.tooltip;
-          item.tooltip = buildPortsTooltip(takenPorts);
+          item.tooltip = buildPortsTooltip(listedPorts, takenPorts);
           if (oldTooltip !== item.tooltip) {
           }
         } else {
@@ -59,11 +68,16 @@ export async function updatePortsStatusBarItem(
         }
         item.show();
       });
-    });
+    }
     if (cycles > 0) {
       setTimeout(async function () {
-        updatePortsStatusBarItem(item, cycles - 1);
+        updatePortsStatusBarItem(item, projectLocalConfig, cycles - 1);
       }, 1000);
+    }
+    else {
+      setTimeout(async function () {
+        updatePortsStatusBarItem(item, projectLocalConfig, cycles);
+      }, 5000);
     }
   } else {
     item.text = "No active ports";
@@ -71,33 +85,43 @@ export async function updatePortsStatusBarItem(
   }
 }
 
-export function getListedPorts(): Array<Number> {
+export function readYaml(path: string): any {
   const workspaceFolder = getMainWorkspace();
   if (!workspaceFolder) {
-    return [];
+    return;
   }
-  let stdout = "";
-  try {
-    stdout = ps
-      .execSync(
-        "find . -maxdepth 3 -type f -name 'local.config.yaml' -exec grep -H 'port:' {} \\;",
-        {
-          cwd: workspaceFolder.uri.path,
-        }
-      )
-      .toString();
-  } catch (e) {}
-  return Array.from(
-    new Set(
-      stdout
-        .split("\n")
-        .filter((l) => l.includes("port:"))
-        .map((l) => parseInt(l.split("port:")[1].trim()))
-    )
-  );
+  path = `${workspaceFolder.uri.path}/${path}`;
+
+  if (!path || !fs.existsSync(path)) {
+    return;
+  }
+  const file = ps.execSync(`cat ${path}`).toString();
+  return YAML.parse(file);
 }
 
-function buildPortsTooltip(takenPorts: Array<Number>): vscode.MarkdownString {
+export function getListedPorts(projectLocalConfig: any): Map<Number, String> {
+  if (!projectLocalConfig || !projectLocalConfig.services) {
+    return new Map();
+  }
+  let ports = new Map();
+  for (let [serviceName, serviceRaw] of Object.entries(
+    projectLocalConfig.services
+  )) {
+    if (!serviceRaw) {
+      continue;
+    }
+    const service = <Service>serviceRaw;
+    if (service.port) {
+      ports.set(service.port, serviceName);
+    }
+  }
+  return ports;
+}
+
+function buildPortsTooltip(
+  listedPorts: Map<Number, String>,
+  takenPorts: Array<Number>
+): vscode.MarkdownString {
   let tooltip = new vscode.MarkdownString();
   tooltip.isTrusted = true;
   tooltip.supportHtml = true;
@@ -106,7 +130,9 @@ function buildPortsTooltip(takenPorts: Array<Number>): vscode.MarkdownString {
     const url = process.env.ENV_URL
       ? `http://${port}.${process.env.ENV_URL}`
       : `http://localhost:${port}`;
-    tooltip.value += `<p><a href='${url}' target='_blank'>${url}</a></p>`;
+    tooltip.value += `<p><a href='${url}' target='_blank'>${listedPorts.get(
+      port
+    )}</a></p>`;
   });
   tooltip.value += "</div>";
   return tooltip;